requirements.txt
Django
pillow
djangorestframework==3.16.0